Refactor inputkriteria_it.html: Update layout and improve form structure for better usability
- Reorganized sections for clarity and consistency.
- Added buttons for selecting data sources (Inventori and Pengadaan).
- Enhanced form fields for project and role selection with improved data handling.
- Updated styling for better visual hierarchy and responsiveness.
- Included summary section to clarify DSS configuration outcomes.
- Implemented JavaScript enhancements for dynamic role filtering and default role settings.

+++ querry database/utilty.sql
@@ -171,8 +171,8 @@ CREATE OR REPLACE FUNCTION GetFilteredLaptopPengadaan(
     f_id_laptop_pengadaan VARCHAR DEFAULT NULL,
 
     f_harga NUMERIC DEFAULT NULL,
-    f_min_harga NUMERIC DEFAULT NULL,
-    f_max_harga NUMERIC DEFAULT NULL,
+    f_min_harga INT DEFAULT NULL,
+    f_max_harga INT DEFAULT NULL,
 
     f_gpu VARCHAR DEFAULT NULL,
 
@@ -206,7 +206,7 @@ CREATE OR REPLACE FUNCTION GetFilteredLaptopPengadaan(
 )
 RETURNS TABLE (
     id_laptop_pengadaan VARCHAR,
-    harga NUMERIC,
+    harga INT,
     gpu VARCHAR,
     ukuran_layar FLOAT,
     baterai FLOAT,
